step tree parser
note: internally, the steps are stored as a DAG. parsing currently only allows
for the definition of a tree instead of a full DAG, but this can be changed in
the future if necessary. this means that currently, it is not possible to use
the output of more than one step as the input of a dependent step.