do not remove ignored tools from comparison
If a task changed it will now list a tool with
retrospectivly ignored tool version as missing.
It would be nice if such a tool would be ignored
to determine if the task has changed but the old
hash of the step, that is used to determine if a
child step is changed, does include the old tool
version anyway and thus a tool version cannot be
ignored retrospectivly entirely. Instead one has
to remove outputs that should be re-run and pass
the --ignore option to ignore all changes when
the pipeline is continued.