move output later
The output files are now moved directly befor creating the annotattion
file. This way the task only becomes complete after the hashing was
completed. A missing annotation file for an otherwise complete task now marks
the task as CHANGED since it is very unlikly to be missing due to an error.
Since CHANGED tasks are not overwritten by default this protects the
output.

Additionally the hashing progress is reported in interactive shells
and on the info logging level.